Jim's Review
Throw unsupported exceptions, rename classes and variables, and regenerate clients.
