+++ tools/openapi.yaml
@@ -212,8 +212,9 @@ paths:
     post:
       tags:
       - Field Data Resource
-      summary: For advanced users who already understand the constraints of your field
-        data and want to create it in one request.
+      summary: "For advanced users, combines the two separate requests of Analyze\
+        \ File and Create From Analyzed File. The following files are accepted: .tif\
+        \ and .zip."
       operationId: analyzeFileAndCreate
       requestBody:
         content:
