+++ webapp-ng/src/app/core/modules/openapi/api/field-data-resource.service.ts
@@ -275,7 +275,7 @@ export class FieldDataResourceService implements FieldDataResourceServiceInterfa
     }
 
     /**
-     * For advanced users who already understand the constraints of your field data and want to create it in one request.
+     * For advanced users, combines the two separate requests of Analyze File and Create From Analyzed File. The following files are accepted: .tif and .zip.
      * @param file 
      * @param fileName 
      * @param extent
